requirements.txt
# requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
PyQt5>=5.15.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
easyocr>=1.7.0
lama-cleaner>=1.2.0
psutil>=5.9.0